+++ source/common/js/colors.js
@@ -290,6 +290,7 @@
         },
 
         expandHexColor: function(hex){
+            console.log(hex);
             if (typeof hex !== "string") {
                 throw new Error("Value is not a string!");
             }
@@ -332,29 +333,29 @@
                 .replace(/[^\d.,]/g, "")
                 .split(",")
                 .map(function(v) {
-                    return _color.includes("hs") ? parseFloat(v) : parseInt(v);
+                    return v.indexOf(".") > -1 ? parseFloat(v) : parseInt(v);
                 });
 
             if (_color[0] === "#") {
                 return this.expandHexColor(_color);
             }
 
-            if (_color.includes("rgba")) {
+            if (_color.indexOf("rgba") > -1) {
                 return new RGBA(a[0], a[1], a[2], a[3]);
             }
-            if (_color.includes("rgb")) {
+            if (_color.indexOf("rgb") > -1) {
                 return new RGB(a[0], a[1], a[2]);
             }
-            if (_color.includes("cmyk")) {
+            if (_color.indexOf("cmyk") > -1) {
                 return new CMYK(a[0], a[1], a[2], a[3]);
             }
-            if (_color.includes("hsv")) {
+            if (_color.indexOf("hsv") > -1) {
                 return new HSV(a[0], a[1], a[2]);
             }
-            if (_color.includes("hsla")) {
+            if (_color.indexOf("hsla") > -1) {
                 return new HSLA(a[0], a[1], a[2], a[3]);
             }
-            if (_color.includes("hsl")) {
+            if (_color.indexOf("hsl") > -1) {
                 return new HSL(a[0], a[1], a[2]);
             }
             return _color;
@@ -1112,17 +1113,18 @@
             return this.createScheme.apply(this, arguments)
         },
 
-        mix: function(color1, color2){
+        mix: function(color1, color2, returnAs){
             var c1 = this.toRGBA(color1);
             var c2 = this.toRGBA(color2);
             var result = new RGBA();
+            var to = (""+returnAs).toLowerCase() || "hex";
 
             result.r = Math.round((c1.r + c2.r) / 2);
             result.g = Math.round((c1.g + c2.g) / 2);
             result.b = Math.round((c1.b + c2.b) / 2);
             result.a = Math.round((c1.a + c2.a) / 2);
 
-            return result;
+            return this["to"+to.toUpperCase()](result);
         }
     };
 
@@ -1133,13 +1135,18 @@
 
     ColorType.prototype = {
         _setValue: function(color){
+            var _color;
+
             if (typeof color === "string") {
-                color = Colors.expandHexColor(Colors.parse(color));
+                _color = Colors.parse(color);
+                _color = typeof _color === "string" ? Colors.expandHexColor(_color) : _color;
             }
-            if (!Colors.isColor(color)) {
-                color = "#000000";
+
+            if (!Colors.isColor(_color)) {
+                _color = "#000000";
             }
-            this._value = color;
+
+            this._value = _color;
             this._type = Colors.colorType(this._value);
         },
 
@@ -1356,14 +1363,22 @@
         },
 
         mix: function(color){
-            var mixedColor = Colors.mix(this._value, color);
-            this._value = Colors['to'+this._type.toUpperCase()](mixedColor);
+            var mixedColor = Colors.mix(this._value, color, this._type);
+            this._value = mixedColor;
             return this;
         }
     }
 
     Metro.colors = Colors.init();
     window.Color = Metro.Color = ColorType;
+    window.ColorPrimitive = Metro.colorPrimitive = {
+        RGB: RGB,
+        RGBA: RGBA,
+        HSV: HSV,
+        HSL: HSL,
+        HSLA: HSLA,
+        CMYK: CMYK
+    };
 
     if (window.METRO_GLOBAL_COMMON === true) {
         window.Colors = Metro.colors;